help KbWait/KbCheck: Fix description of 'deviceNumber'.
It is not true that the default keyboard device chosen in absence
of the optional 'deviceNumber' is always the 1st USB HID keyboard.

That was true in the better days of Apples OSX, but not neccessarily
since we had to work around HID bugs in the trainwreck.

It was never true for Linux or Windows. Formulate the help text
more fuzzy to account for this.